Add test cases for junit5
