+++ kafka-junit5/src/test/java/com/salesforce/kafka/test/junit5/SharedKafkaTestResourceWithSaslPlainTest.java
@@ -23,23 +23,22 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.kafka.test.junit4;
+package com.salesforce.kafka.test.junit5;
 
 import com.salesforce.kafka.test.KafkaTestUtils;
-import com.salesforce.kafka.test.listeners.SaslSslListener;
-import org.junit.ClassRule;
+import com.salesforce.kafka.test.listeners.SaslPlainListener;
+import org.junit.jupiter.api.extension.RegisterExtension;
 
 /**
  * NOTE: This test case assumes you've started the JVM with the argument.
  *
  * -Djava.security.auth.login.config=kafka-junit-core/src/test/resources/jaas.conf
  *
- * and run the script/generateCertificatesForTest.sh script.
- *
- * Runs smoke tests against an SASL+SSL enabled cluster.
+ * Runs smoke tests against an SASL enabled cluster.
  * @see AbstractSharedKafkaTestResourceTest for test case definitions.
  */
-public class SharedKafkaTestResourceWithSaslSslTest extends AbstractSharedKafkaTestResourceTest {
+public class SharedKafkaTestResourceWithSaslPlainTest extends AbstractSharedKafkaTestResourceTest {
+
     /**
      * We have a two node kafka cluster that gets started when this test class is initialized.
      *
@@ -49,24 +48,18 @@ public class SharedKafkaTestResourceWithSaslSslTest extends AbstractSharedKafkaT
      * This example we start a cluster with
      *  - 2 brokers (defaults to a single broker)
      *  - configure the brokers to disable topic auto-creation.
-     *  - Enable SASL_SSL authentication, using username 'kafkaclient' and password 'client-secret'
+     *  - Enable SASL_PLAIN authentication, using username 'kafkaclient' and password 'client-secret'
      */
-    @ClassRule
-    public static final SharedKafkaTestResource sharedKafkaTestResource = new SharedKafkaTestResource()
+    @RegisterExtension
+    static final SharedKafkaTestResource sharedKafkaTestResource = new SharedKafkaTestResource()
         // Start a cluster with 2 brokers.
         .withBrokers(2)
         // Disable topic auto-creation.
         .withBrokerProperty("auto.create.topics.enable", "false")
         // Register and configure SASL PLAIN authentication on cluster.
-        .registerListener(new SaslSslListener()
+        .registerListener(new SaslPlainListener()
         .withUsername("kafkaclient")
         .withPassword("client-secret")
-        .withClientAuthRequested()
-        .withKeyStoreLocation(SharedKafkaTestResourceWithSaslSslTest.class.getClassLoader().getResource("kafka.keystore.jks").getFile())
-        .withKeyStorePassword("password")
-        .withTrustStoreLocation(SharedKafkaTestResourceWithSaslSslTest.class.getClassLoader().getResource("kafka.truststore.jks").getFile())
-        .withTrustStorePassword("password")
-        .withKeyPassword("password")
     );
 
     /**
