+++ kafka-junit5/src/test/java/com/salesforce/kafka/test/junit5/SharedKafkaTestResourceWithSslTest.java
@@ -23,25 +23,21 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.kafka.test.junit4;
+package com.salesforce.kafka.test.junit5;
 
 import com.salesforce.kafka.test.KafkaTestUtils;
-import com.salesforce.kafka.test.listeners.SaslSslListener;
-import org.junit.ClassRule;
+import com.salesforce.kafka.test.listeners.SslListener;
+import org.junit.jupiter.api.extension.RegisterExtension;
 
 /**
- * NOTE: This test case assumes you've started the JVM with the argument.
+ * NOTE: This test assumes you've run the script/generateCertificatesForTest.sh script.
  *
- * -Djava.security.auth.login.config=kafka-junit-core/src/test/resources/jaas.conf
- *
- * and run the script/generateCertificatesForTest.sh script.
- *
- * Runs smoke tests against an SASL+SSL enabled cluster.
+ * Runs smoke tests against an SSL enabled cluster.
  * @see AbstractSharedKafkaTestResourceTest for test case definitions.
  */
-public class SharedKafkaTestResourceWithSaslSslTest extends AbstractSharedKafkaTestResourceTest {
+class SharedKafkaTestResourceWithSslTest extends AbstractSharedKafkaTestResourceTest {
     /**
-     * We have a two node kafka cluster that gets started when this test class is initialized.
+     * We have two node embedded kafka cluster that gets started when this test class is initialized.
      *
      * It's automatically started before any methods are run via the @ClassRule annotation.
      * It's automatically stopped after all of the tests are completed via the @ClassRule annotation.
@@ -49,22 +45,20 @@ public class SharedKafkaTestResourceWithSaslSslTest extends AbstractSharedKafkaT
      * This example we start a cluster with
      *  - 2 brokers (defaults to a single broker)
      *  - configure the brokers to disable topic auto-creation.
-     *  - Enable SASL_SSL authentication, using username 'kafkaclient' and password 'client-secret'
+     *  - enables SSL authentication with a test/dummy key and trust stores.
      */
-    @ClassRule
-    public static final SharedKafkaTestResource sharedKafkaTestResource = new SharedKafkaTestResource()
+    @RegisterExtension
+    static final SharedKafkaTestResource sharedKafkaTestResource = new SharedKafkaTestResource()
         // Start a cluster with 2 brokers.
         .withBrokers(2)
         // Disable topic auto-creation.
         .withBrokerProperty("auto.create.topics.enable", "false")
-        // Register and configure SASL PLAIN authentication on cluster.
-        .registerListener(new SaslSslListener()
-        .withUsername("kafkaclient")
-        .withPassword("client-secret")
+        // Register and configure SSL authentication on cluster.
+        .registerListener(new SslListener()
         .withClientAuthRequested()
-        .withKeyStoreLocation(SharedKafkaTestResourceWithSaslSslTest.class.getClassLoader().getResource("kafka.keystore.jks").getFile())
+        .withKeyStoreLocation(SharedKafkaTestResourceWithSslTest.class.getClassLoader().getResource("kafka.keystore.jks").getFile())
         .withKeyStorePassword("password")
-        .withTrustStoreLocation(SharedKafkaTestResourceWithSaslSslTest.class.getClassLoader().getResource("kafka.truststore.jks").getFile())
+        .withTrustStoreLocation(SharedKafkaTestResourceWithSslTest.class.getClassLoader().getResource("kafka.truststore.jks").getFile())
         .withTrustStorePassword("password")
         .withKeyPassword("password")
     );
